+++ src/Doctrine/ArrayNormalizableThroughLookupType.php
@@ -8,24 +8,21 @@ use DigitalCraftsman\SelfAwareNormalizers\Serializer\ArrayNormalizable;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
 use Doctrine\DBAL\Types\Type;
 
-/**
- * @deprecated Will be removed in version 2.0.0. Should be replaced with automatic registration.
- */
-abstract class ArrayNormalizableType extends Type
+final class ArrayNormalizableThroughLookupType extends Type
 {
-    abstract public static function getTypeName(): string;
-
-    /**
-     * @return class-string<ArrayNormalizable>
-     */
-    abstract public static function getClass(): string;
-
-    /**
-     * @codeCoverageIgnore
-     */
     #[\Override]
     public function getSQLDeclaration(array $column, AbstractPlatform $platform): string
     {
+        $className = $this->getClassNameThroughLookup();
+        $reflectionClass = new \ReflectionClass($className);
+
+        if ($reflectionClass->implementsInterface(NormalizableTypeWithSQLDeclaration::class)) {
+            /**
+             * @var NormalizableTypeWithSQLDeclaration $className
+             */
+            return $className::getSQLDeclaration($column, $platform);
+        }
+
         return $platform->getJsonbTypeDeclarationSQL($column);
     }
 
@@ -39,12 +36,11 @@ abstract class ArrayNormalizableType extends Type
             return null;
         }
 
-        /** @var class-string<ArrayNormalizable> $class */
-        $class = static::getClass();
+        $className = $this->getClassNameThroughLookup();
 
         $array = json_decode($value, true, 512, JSON_THROW_ON_ERROR);
 
-        return $class::denormalize($array);
+        return $className::denormalize($array);
     }
 
     /**
@@ -60,5 +56,21 @@ abstract class ArrayNormalizableType extends Type
         $array = $value->normalize();
 
         return json_encode($array, JSON_THROW_ON_ERROR);
+    }
+
+    /**
+     * @return class-string<ArrayNormalizable>
+     */
+    private function getClassNameThroughLookup(): string
+    {
+        /**
+         * @var class-string<ArrayNormalizable> $className
+         */
+        $className = self::lookupName($this);
+        if (!class_exists($className)) {
+            throw new Exception\InvalidTypeName($className);
+        }
+
+        return $className;
     }
 }