+++ src/Doctrine/StringEnumType.php
@@ -7,6 +7,9 @@ namespace DigitalCraftsman\SelfAwareNormalizers\Doctrine;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
 use Doctrine\DBAL\Types\Type;
 
+/**
+ * @deprecated Will be removed in version 2.0.0. Should be replaced with automatic registration.
+ */
 abstract class StringEnumType extends Type
 {
     abstract public static function getTypeName(): string;