+++ src/Doctrine/StringNormalizableTypeAsTypeText.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigitalCraftsman\SelfAwareNormalizers\Doctrine;
+
+interface StringNormalizableTypeAsTypeText
+{
+}