+++ tests/Test/ValueObject/UserRole.php
@@ -4,8 +4,13 @@ declare(strict_types=1);
 
 namespace DigitalCraftsman\SelfAwareNormalizers\Test\ValueObject;
 
-enum UserRole: string
+use DigitalCraftsman\SelfAwareNormalizers\Serializer\StringNormalizable;
+use DigitalCraftsman\SelfAwareNormalizers\Serializer\StringNormalizableEnumTrait;
+
+enum UserRole: string implements StringNormalizable
 {
+    use StringNormalizableEnumTrait;
+
     case ROLE_USER = 'ROLE_USER';
     case ROLE_ADMIN = 'ROLE_ADMIN';
 }